refactor: migrate to framework-neutral architecture and introduce UI loader
