+++ ui/web/main/desktop/app.js
@@ -1,231 +1,16 @@
 //---------------------------------------------------------------------
-// Application UI
+// Application UI (build entry point)
 //---------------------------------------------------------------------
 // Copyright (C) 2007-2025 The NOC Project
 // See LICENSE for details
 //---------------------------------------------------------------------
+// This file is the esbuild entry point. NocLoaderPlugin prepends an import for
+// every Ext.define class (topologically sorted) ahead of this code, so all NOC
+// classes are defined before the loader runs. The application itself is started
+// by the framework-neutral loader chain (index -> bootstrap -> ui-loader ->
+// ext-application); ExtJS is no longer the root. See loader/index.js and
+// ui/docs/extjs-migration-summary.md.
+//---------------------------------------------------------------------
+import "./loader/index.js";
+
 console.debug("Defining NOC application");
-
-Ext.application({
-  name: "NOC",
-  paths: {
-    "NOC": "/ui/web",
-    "Ext.ux": "/ui/web/ux",
-  },
-
-  requires: [
-    "NOC.main.desktop.Application",
-    "NOC.main.desktop.LoginView",
-  ],
-
-  launch: function(){
-    Ext.Ajax.request({
-      url: "/api/login/is_logged/",
-      method: "GET",
-      scope: this,
-      success: function(response){
-        if(Ext.decode(response.responseText)){
-          this.openApplication();
-          return;
-        }
-
-        var hash = location.hash,
-          newUrl = location.protocol + "//" + 
-                      location.host + 
-            "?uri=/" + hash;
-
-        history.replaceState({}, document.title, newUrl);
-        this.openLogin();
-      },
-      failure: function(response){
-        console.error("Request failed:", response.status);
-      },
-    });
-  },
-  openApplication: function(){
-    Ext.setGlyphFontFamily("FontAwesome");
-    console.log("Initializing navigation");
-    // Navigation API migration (Phase 1, hash-based URLs). This replaces the
-    // former Ext.History facade; see ui/docs/navigation-api-migration.md
-    NOC.core.Navigation.init();
-    console.log("NOC application starting");
-    this.settings();
-  },
-  openLogin: function(){
-    console.log("NOC login starting");
-    Ext.create("NOC.main.desktop.LoginView", {
-      listeners: {
-        scope: this,
-        afterRender: this.hideSplashScreen,
-      },
-    });
-    console.log("NOC login started");
-  },
-  settings: function(){
-    Ext.Ajax.request({
-      url: "/main/desktop/settings/",
-      method: "GET",
-      scope: this,
-      success: function(response){
-        var setup = Ext.decode(response.responseText);
-        console.log("!!!");
-        console.log("!!! Running NOC desktop");
-        console.log("!!!");
-        // Initialize loader
-        Ext.BLANK_IMAGE_URL = "/ui/web/img/s.gif";
-        NOC.settings = {
-          systemId: setup.system_uuid ? setup.system_uuid : null,
-          brand: setup.brand,
-          features: setup.features || [],
-          installation_name: setup.installation_name,
-          preview_theme: setup.preview_theme,
-          language: setup.language,
-          branding_color: setup.branding_color,
-          branding_background_color: setup.branding_background_color,
-          enable_search: setup.enable_search,
-          gitlab_url: setup.gitlab_url,
-          collections: {
-            allow_sharing: setup.collections.allow_sharing,
-            allow_overwrite: setup.collections.allow_overwrite,
-            project_id: setup.collections.project_id,
-          },
-          gis: {
-            base: {
-              enable_blank: setup.gis.base.enable_blank,
-              "enable_osm": setup.gis.base.enable_osm,
-              "enable_google_roadmap": setup.gis.base.enable_google_roadmap,
-              "enable_google_hybrid": setup.gis.base.enable_google_hybrid,
-              "enable_google_sat": setup.gis.base.enable_google_sat,
-              "enable_google_terrain": setup.gis.base.enable_google_terrain,
-              "enable_tile1": setup.gis.base.enable_tile1,
-              "enable_tile2": setup.gis.base.enable_tile2,
-              "enable_tile3": setup.gis.base.enable_tile3,
-              "enable_yandex_roadmap": setup.gis.base.enable_yandex_roadmap,
-              "enable_yandex_hybrid": setup.gis.base.enable_yandex_hybrid,
-              "enable_yandex_sat": setup.gis.base.enable_yandex_sat,
-            },
-            custom: {
-              tile1: {
-                name: setup.gis.custom.tile1.name,
-                url: setup.gis.custom.tile1.url,
-                subdomains: setup.gis.custom.tile1.subdomains,
-              },
-              tile2: {
-                name: setup.gis.custom.tile2.name,
-                url: setup.gis.custom.tile2.url,
-                subdomains: setup.gis.custom.tile2.subdomains,
-              },
-              tile3: {
-                name: setup.gis.custom.tile3.name,
-                url: setup.gis.custom.tile3.url,
-                subdomains: setup.gis.custom.tile3.subdomains,
-              },
-            },
-            "yandex_supported": setup.gis.yandex_supported,
-            default_layer: setup.gis.default_layer,
-          },
-          traceExtJSEvents: setup.traceExtJSEvents,
-          helpUrl: setup.helpUrl,
-          helpBranch: setup.helpBranch,
-          helpLanguage: setup.helpLanguage,
-          timezone: setup.timezone,
-          enable_remote_system_last_extract_info: setup.enable_remote_system_last_extract_info,
-          enableHelp: setup.helpUrl && setup.helpUrl !== "",
-          has_geocoder: setup.has_geocoder,
-        };
-        NOC.templates = {};
-        // Change title
-        document.title = setup.brand + "|" + setup.installation_name;
-        this.createCss(setup.color_scheme || {});
-        this.app = Ext.create("NOC.main.desktop.Application", {
-          listeners: {
-            scope: this,
-            applicationReady: this.hideSplashScreen,
-          },
-        });
-      },
-      failure: function(response){
-        console.error("Request failed:", response.status);
-      },
-    });
-  },
-  hideSplashScreen: function(){
-    var mask = Ext.get("noc-loading-mask"),
-      parent = Ext.get("noc-loading");
-    mask.fadeOut({
-      callback: function(){
-        mask.destroy();
-      },
-    });
-    parent.fadeOut({
-      callback: function(){
-        parent.destroy();
-      },
-    });
-  },
-  createCss: function(scheme){
-    var classes = scheme.style || [],
-      style = document.createElement("style"),
-      root = document.documentElement,
-      css = "";
-    classes.forEach(cl => {
-      css += `.${cl.name}, .${cl.name} td{`;
-      if(!(Object.hasOwn(cl, "fontWeight") || Object.hasOwn(cl, "font-weight"))){
-        cl.fontWeight = "normal";
-      }
-      if(!(Object.hasOwn(cl, "fontStyle") || Object.hasOwn(cl, "font-style"))){
-        cl.fontStyle = "normal";
-      }
-      if(!(Object.hasOwn(cl, "textDecoration") || Object.hasOwn(cl, "text-decoration"))){
-        cl.textDecoration = "none";
-      }
-      Object.entries(cl)
-        .filter((name) => name[0] !== "name")
-        .forEach(([key, value]) => {
-          switch(key){
-            case "color":
-              css += `color: var(--${cl.name});`;
-              root.style.setProperty(`--${cl.name}`, value);
-              break;
-            case "background-color":
-            case "backgroundColor": {
-              let v = `${cl.name}-bg`;
-              css += `background-color: var(--${v});`;
-              root.style.setProperty(`--${v}`, value);
-              break;
-            }
-            case "font-weight":
-            case "fontWeight": {
-              let v = `${cl.name}-fw`;
-              css += `font-weight: var(--${v});`;
-              root.style.setProperty(`--${v}`, value);
-              break;
-            }
-            case "font-style":
-            case "fontStyle": {
-              let v = `${cl.name}-fs`;
-              css += `font-style: var(--${v});`;
-              root.style.setProperty(`--${v}`, value);
-              break;
-            }
-            case "text-decoration":
-            case "textDecoration": {
-              let v = `${cl.name}-td`;
-              css += `text-decoration: var(--${v});`;
-              root.style.setProperty(`--${v}`, value);
-              break;
-            }
-            default:
-              css += `${key}: ${value}; `;
-          }
-        });
-      css += "} ";
-    });
-    if(style.styleSheet){
-      style.styleSheet.cssText = css;
-    } else{
-      style.appendChild(document.createTextNode(css));
-    }
-    document.getElementsByTagName("head")[0].appendChild(style);
-  },
-});
